hasher_t: Replace use of MD4 with XXH3_128
xxHash is notably faster than MD4. The latter was chosen primarily
for speed, so this change is in line with the BuildCache design.

After this change, old cached items from previous MD4-based versions
of BuildCache will no longer produce cache hits (so the effect
should be similar to clearing the cache).