ti_common_wrapper: Read also fancy response files
For the compilers from TI, handle also response files with CRLF as line
endings. The logs are very hard to interpret without this fix.

Ignore lines beginning with a '#', because those should be treated as
comments according to the documentation from TI.

If a /* C style comment */ is detected, throw an exception because those
are (still) unsupported by BuildCache.