CI: Only release a single Windows package
There is no need to release both MinGW and MSVC packages, as it is
mostly confusing for users (what is the difference between the two
packages?).

Instead we release the MSVC package as "buildcache-windows.zip", but
we still test-build with MinGW (e.g. we get clang-tidy linting for
the Windows code).